+++ src/lib.rs
@@ -12,6 +12,7 @@ pub mod players_manager;
 pub mod powers;
 pub mod stats;
 pub mod target;
+pub mod testing_all_characters;
 pub mod testing_atk;
 pub mod testing_effect;
 pub mod testing_target;
