+++ src/testing_target.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use crate::target::TargetInfo;
 
 pub fn build_target_boss_indiv() -> TargetInfo {
